JI-6710 Fix increased performance
This is done by only processing relevant elements and at a shorter
interval, as opposed to the whole page every 0.5s. In general,
the performance increase is at least >10x, but much more for
larger content with loads of formulas. Also, we've removed old
observer solutions, now we only rely on modern mutations.