JI-6710 optimize update function calls by calling it less often (#25)
* JI-6710 optimize update function calls by calling it less often

* JI-6710 Fix increased performance

This is done by only processing relevant elements and at a shorter
interval, as opposed to the whole page every 0.5s. In general,
the performance increase is at least >10x, but much more for
larger content with loads of formulas. Also, we've removed old
observer solutions, now we only rely on modern mutations.

* JI-6710 Fix auto-merge issues + processing of entire body

---------

Co-authored-by: Frode Petterson <frode.andre.petterson@gmail.com>